进行了一番重构，解决 shell 文件过长过大的问题
